Code refactoring and bug fix
- fix test manifests location
- fix file location in constant
- change e2e package name
