+++ hcsctl/e2e/cdi/cdi_test.go
@@ -1,4 +1,4 @@
-package testCDI
+package cdi
 
 import (
 	. "github.com/onsi/ginkgo"
