+++ hcsctl/e2e/cdi/ginkgotest_suite_test.go
@@ -1,4 +1,4 @@
-package testCDI
+package cdi
 
 import (
 	"testing"
