+++ hcsctl/e2e/cdi/helper.go
@@ -1,4 +1,4 @@
-package testCDI
+package cdi
 
 import (
 	"flag"
