--- hcsctl/e2e/ceph/constant.go
@@ -1,9 +1,0 @@
-package testCeph
-
-import "time"
-
-const (
-	PollingIntervalDefault              = 3 * time.Second
-	TimeoutForDeletingNamespace         = 500 * time.Second *2
-	PollingIntervalForDeletingNamespace = 10 * time.Second
-)
