--- hcsctl/e2e/ceph/manifests/cephfs-pvc.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: PersistentVolumeClaim
-metadata:
-  name: cephfs-pvc
-  namespace: test-namespace
-spec:
-  accessModes:
-  - ReadWriteMany
-  resources:
-    requests:
-      storage: 100Mi
-  storageClassName: csi-cephfs-sc
