--- hcsctl/e2e/ceph/manifests/namespace.yaml
@@ -1,4 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: test-namespace
