--- hcsctl/e2e/ceph/manifests/rbd-pvc.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: PersistentVolumeClaim
-metadata:
-  name: rbd-pvc
-  namespace: test-namespace
-spec:
-  accessModes:
-  - ReadWriteOnce
-  resources:
-    requests:
-      storage: 100Mi
-  storageClassName: rook-ceph-block
