+++ hcsctl/e2e/rook/go.mod
@@ -1,0 +1,14 @@
+module tmax.com/hypercloud-storage/ginkgotest
+
+go 1.13
+
+require (
+	github.com/coreos/go-systemd v0.0.0-20191104093116-d3cd4ed1dbcf // indirect
+	github.com/coreos/pkg v0.0.0-20180928190104-399ea9e2e55f // indirect
+	github.com/onsi/ginkgo v1.11.0
+	github.com/onsi/gomega v1.7.0
+	github.com/rook/rook v1.2.3
+	k8s.io/api v0.0.0-20200812052025-e70c4f3b2093
+	k8s.io/apimachinery v0.0.0-20200812051649-2a282836017b
+	k8s.io/client-go v0.0.0-20200812053003-c8494005d10a
+)
